Organise tests into test suite per method
